requirements.txt
torch
torchvision
opencv-python
albumentations
fastapi
uvicorn
python-multipart
scikit-learn
pandas
numpy
matplotlib
pyyaml
requests
tqdm
jupyter
notebook
bing-image-downloader
fastdownload
fastcore
seaborn
